refactor(UIFactory/default-ui-facotry): should return React.lazy versions of components [YTFRONT-3814]
