+++ packages/ui/src/ui/containers/ACL/SubjectsControl/lazy.tsx
@@ -1,0 +1,8 @@
+import React from 'react';
+import withLazyLoading from '../../../hocs/withLazyLoading';
+
+export const YTSubjectSuggestLazy = withLazyLoading(
+    React.lazy(async () => {
+        return {default: (await import('./YTSubjectSuggest')).YTSubjectSuggest};
+    }),
+);
