+++ packages/ui/src/ui/containers/UserSuggest/YTUserSuggest.tsx
@@ -3,7 +3,7 @@ import {useSelector} from 'react-redux';
 
 import {getAllUserNames} from '../../store/selectors/global';
 import Select from '../../components/Select/Select';
-import {SubjectName} from '../../components/SubjectLink/SubjectLink';
+import {SubjectName} from '../../components/SubjectLink/DefaultSubjectLink';
 import {useAllUserNamesFiltered} from '../../hooks/global';
 import {UserSuggestProps} from './UserSuggest';
