+++ packages/ui/src/ui/containers/UserSuggest/YTUserSuggestLazy.tsx
@@ -1,0 +1,8 @@
+import React from 'react';
+import withLazyLoading from '../../hocs/withLazyLoading';
+
+export const YTUserSuggestLazy = withLazyLoading(
+    React.lazy(async () => {
+        return {default: (await import('./YTUserSuggest')).YTUserSuggest};
+    }),
+);
