+++ packages/ui/src/ui/pages/operations/OperationDetail/tabs/monitor/lazy.tsx
@@ -1,0 +1,6 @@
+import withLazyLoading from '../../../../../hocs/withLazyLoading';
+import React from 'react';
+
+export const OperationDetailsMonitorLinksLazy = withLazyLoading(
+    React.lazy(async () => await import('./OperationDetailsMonitorLinks')),
+);
