+++ packages/ui/src/ui/pages/query-tracker/QueryResultsVisualization/index.tsx
@@ -1,7 +1,17 @@
 import React from 'react';
-import {QueryResultsVisualization} from './containers/QueryResultsVisualization/QueryResultsVisualization';
+import withLazyLoading from '../../../hocs/withLazyLoading';
+
+const QueryResultsVisualizationLazy = withLazyLoading(
+    React.lazy(async () => {
+        return {
+            default: (
+                await import('./containers/QueryResultsVisualization/QueryResultsVisualization')
+            ).QueryResultsVisualization,
+        };
+    }),
+);
 
 export const CUSTOM_QUERY_REQULT_TAB = {
     title: 'Chart',
-    renderContent: (props: any) => <QueryResultsVisualization {...props} />,
+    renderContent: (props: any) => <QueryResultsVisualizationLazy {...props} />,
 };
